requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-rest==0.8.7
djangorestframework==3.14.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
